pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "substation"
dynamic = ["version"]
description = "A versatile SDR band scanner for RTL-SDR, HackRF, AirSpy, and SoapySDR devices"
authors = [
    { name = "Simon Holliday" }
]
license = { text = "AGPL-3.0" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pydantic>=2.0",
    "pyrtlsdr>=0.3.0,<0.4.0",  # 0.4.0 requires rtlsdr_set_dithering (keenerd fork only)
    "python_hackrf",
    "noisereduce>=3.0.3",
    "PyYAML>=6.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.1",
    # SoapySDR (for AirSpy and other devices) is a system package — install
    # via apt (python3-soapysdr) and use --system-site-packages on the venv.
    # See installation.txt section 4.
]

[project.scripts]
substation = "substation.cli:main"

[tool.setuptools.packages.find]
include = ["substation*"]

[tool.setuptools.dynamic]
version = {attr = "substation.__version__"}

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
